video-service/requirements.txt
opencv-python==4.11.0.86
numpy==1.24.4
redis
uuid
redis
Pillow
python-dotenv
requests>=2.26.0
urllib3>=1.26.14
charset-normalizer>=2.0.0